backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "IPID document parsing"
authors = ["Robin Beaudet <robin.beaudet@zelros.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.16.0"
python-dotenv = "^0.19.2"
PyMuPDF = "^1.19.2"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
